pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "next-whistle-streaming-companion"
version = "0.1.0"
description = "Roller derby live streaming companion service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "obsws-python>=1.7",
    "httpx>=0.27",
    "websockets>=12.0",
    "jinja2>=3.1",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.scripts]
nwsc = "nwsc.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/nwsc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
